Fix Swiftlint warnings
